improve KGquery and work on axioms editor
