+++ public/vocables/index.html
@@ -65,6 +65,9 @@
 
         <script type="module" src="modules/tools/lineage/lineage_upperOntologies.js"></script>
 
+       <!-- <script type="module" src="modules/tools/containers/Containers_UI.js"></script>
+        <script type="module" src="modules/tools/containers/lineage_containers_treeWidget.js"></script>
+        <script type="module" src="modules/tools/containers/containers_query.js"></script>-->
         <script type="module" src="modules/tools/lineage/lineage_containers.js"></script>
 
         <script type="module" src="modules/tools/lineage/lineage_selection.js"></script>
