+++ public/vocables/modules/tools/containers/Containers_UI.js
@@ -1,5 +1,5 @@
-import Lineage_whiteboard from "./lineage_whiteboard.js";
-import Lineage_styles from "./lineage_styles.js";
+import Lineage_whiteboard from "../lineage/lineage_whiteboard.js";
+import Lineage_styles from "../lineage/lineage_styles.js";
 import Sparql_common from "../../sparqlProxies/sparql_common.js";
 import common from "../../shared/common.js";
 import Sparql_proxy from "../../sparqlProxies/sparql_proxy.js";
@@ -8,93 +8,9 @@ import JstreeWidget from "../../uiWidgets/jstreeWidget.js";
 
 self.lineageVisjsGraph;
 
-var Lineage_containers = (function() {
+var Containers_UI = (function() {
     var self = {};
-    self.containerEdgeColor="#ecba4c"
-    self.containerStyle = { shape: "square", color: "#ecba4c" };
-    self.nbOfancestors_tree_search = 0;
-    self.ancestors_tree_search_areRunning = [false, false, false];
-    self.flag_search = false;
-    self.flag_search_launch_search = false;
-    self.flag_function = function() {
-        self.flag_search = false;
-        if (self.flag_search_launch_search == true) {
-            Lineage_containers.search();
-            self.flag_search_launch_search = false;
-        }
-    };
-    self.add_supplementary_layer_for_types_icon = null;
 
-    self.getContextJstreeMenu = function() {
-        var items = {};
-        items["NodeInfos"] = {
-            label: "Node infos",
-            action: function(_e) {
-                NodeInfosWidget.showNodeInfos(Lineage_sources.activeSource, self.currentContainer, "mainDialogDiv");
-            }
-        };
-        items["GraphNode"] = {
-            label: "Graph node",
-            action: function(_e) {
-                if (self.currentContainer.data.type == "container") {
-                    Lineage_containers.graphResources(Lineage_sources.activeSource, self.currentContainer.data, { onlyOneLevel: true });
-                } else {
-                    Lineage_whiteboard.drawNodesAndParents(self.currentContainer, 0);
-                }
-            }
-        };
-        items["Open node"] = {
-            label: "Open node",
-            action: function(_e) {
-                // $("#lineage_containers_containersJstree").jstree().open_all(self.currentContainer.id);
-                Lineage_containers.listContainerResources(Lineage_sources.activeSource, self.currentContainer, { onlyOneLevel: true, leaves: true });
-            }
-        };
-        items.copyNodes = {
-            label: "Copy Node(s)",
-            action: function(e) {
-                // pb avec source
-                Lineage_whiteboard.copyNode(e);
-                var selectedNodes = $("#lineage_containers_containersJstree").jstree().get_selected(true);
-                Lineage_common.copyNodeToClipboard(selectedNodes);
-            }
-        };
-        items["AddGraphNode"] = {
-            label: "Add selected node to container",
-            action: function(_e) {
-                var graphNodeData = Lineage_whiteboard.currentGraphNode.data;
-                Lineage_containers.addResourcesToContainer(Lineage_sources.activeSource, self.currentContainer, graphNodeData);
-            }
-        };
-        items["PasteNodesInContainer"] = {
-            label: "Paste nodes in container",
-            action: function(_e) {
-                Lineage_containers.pasteNodesInContainer(Lineage_sources.activeSource, self.currentContainer);
-            }
-        };
-
-        items["DeleteContainer"] = {
-            label: "Delete container",
-            action: function(_e) {
-                Lineage_containers.deleteContainer(Lineage_sources.activeSource, self.currentContainer);
-            }
-        };
-
-        items["GraphContainerDescendant"] = {
-            label: "Graph  descendants",
-            action: function(_e) {
-                Lineage_containers.graphResources(Lineage_sources.activeSource, self.currentContainer.data, { descendants: true });
-            }
-        };
-        items["GraphContainerDescendantAndLeaves"] = {
-            label: "Graph  descendants + leaves",
-            action: function(_e) {
-                Lineage_containers.graphResources(Lineage_sources.activeSource, self.currentContainer.data, { leaves: true });
-            }
-        };
-
-        return items;
-    };
 
     self.search = function(memberType, callback) {
         if (!callback) {
@@ -103,38 +19,39 @@ var Lineage_containers = (function() {
         }
 
         var term = $("#Lineage_containers_searchInput").val();
-        var searchWhat = $("#Lineage_containers_searchWhatInput").val();
         var source = Lineage_sources.activeSource;
-
-        if (!memberType) {
-            memberType = "";
-        }
 
         var filter = "";
         if (term) {
             filter = Sparql_common.setFilter("searchValue", null, term);
+        }else{
+            self.sparql_queries.getTopContainers(source, function(err, result) {
+                if (err) {
+                    return callback(err);
+                }
+                if (!term && result.results.binding.length==0) {
+                    return callback("no root container found, cannot display containers hierarchy");
+                }
+
+                result.results.bindings.forEach(function(item) {
+                    rootNodes.push({ id: item.member.value, label: item.memberLabel.value });
+                });
+                callbackSeries();
+            });
+
+
+
         }
 
-        var search_on_container = "";
 
-        if (searchWhat == "current") {
-            if (!self.currentContainer) {
-                alert("no selected container");
-                return;
-            }
-            search_on_container = self.currentContainer.data.id;
-            if (!(self.currentContainer.data.type == "container" || self.currentContainer.data.type.includes("http://www.w3.org/1999/02/22-rdf-syntax-ns#Bag"))) {
-                alert("not a bag");
-                return;
-            }
-        }
 
-        self.currentContainer = null;
+
 
         if ($("#lineage_containers_containersJstree").jstree) {
             $("#lineage_containers_containersJstree").empty();
         }
         var options={}
+        options.depth=1
         if (self.flag_search == false) {
             if (term != "") {
                 options.depth=5
@@ -150,214 +67,172 @@ var Lineage_containers = (function() {
         }
     };
 
-    self.drawContainerJstree = function(source, filter, jstreeDiv, search_on_container, memberType, options, callback) {
-        if (!options) {
-            options = {};
+    /**
+     *
+     * add nodes to a container(owl:bag)
+     *
+     *
+     * @param source
+     * @param container
+     * @param nodesData
+     * @param drawMembershipEdge add the edge (and the node) on the vizGraph
+     */
+    self.addResourcesToContainer = function(source, container, nodesData, drawMembershipEdge, callback) {
+        if (!(container.data.type.includes("http://www.w3.org/1999/02/22-rdf-syntax-ns#Bag") || container.data.type == "container")) {
+            return alert("can only add resources to containers");
         }
-        if (!options.depth) {
-            options.depth = 1;
+        // self.currentContainer=null;
+        if (!Array.isArray(nodesData)) {
+            nodesData = [nodesData];
         }
-        if (!options.filter) {
-            options.filter = "";
-        }
-        options.filter += filter || "";
 
-        var data = [];
-        var rootNodes = [];
-        async.series(
-            [
-                function(callbackSeries) {
-                    if (search_on_container) {
-                        if (!Array.isArray(search_on_container)) {
-                            search_on_container = [search_on_container];
-                        }
-                        return callbackSeries();
-                    }
-                    // determine top container with members if no search_on_container
-                    self.sparql_queries.getTopContainers(source, function(err, result) {
-                        if (err) {
-                            return callback(err);
-                        }
-                        if (err) {
-                            return callback(err);
-                        }
-                        search_on_container = [];
-
-                        result.results.bindings.forEach(function(item) {
-                            search_on_container.push(item.member.value);
-                            rootNodes.push({ id: item.member.value, label: item.memberLabel.value });
-                        });
-                        callbackSeries();
-                    });
-                },
-                // determine top container without  members if no search_on_container and no  container with members
-                function(callbackSeries) {
-                    callbackSeries();
-                },
-                //prepare parents and members
-                function(callbackSeries) {
-                    self.sparql_queries.getContainerDescendants(source, search_on_container, options, function(err, result) {
-                        if (err) {
-                            return callback(err);
-                        }
-                        data = result.results.bindings;
-                        callbackSeries();
-                    });
-                },
-
-                function(callbackSeries) {
-                    var jstreeData = [];
-                    var existingIds = {};
-                    var existingNodes = {};
-
-                    // set rootnodes
-                    rootNodes.forEach(function(item) {
-                        var id = item.id;
-                        var jstreeId = "_" + common.getRandomHexaId(5);
-                        if (!existingIds[id]) {
-                            existingIds[id] = jstreeId;
-                        }
-
-                        //existingIds[id].push(jstreeId);
-
-                        if (!existingNodes[jstreeId]) {
-                            existingNodes[jstreeId] = 1;
-                        }
-                        var node = {
-                            id: existingIds[id],
-                            text: item.label,
-                            parent: "#",
-                            type: "Container",
-                            data: {
-                                type: "Container",
-                                source: source,
-                                id: id,
-                                label: item.label,
-                                parent: "#"
-                                //tabId: options.tabId,
-                            }
-                        };
-
-                        jstreeData.push(node);
-                    });
-
-                    data.forEach(function(item) {
-                        var parentId = item.parent.value;
-                        var types = item.memberTypes.value.split(",");
-
-                        var type = JstreeWidget.selectTypeForIconsJstree(types);
-                        /*
-                        var type = "Container";
-                        if (item.memberTypes.value.indexOf("Class") > 0) {
-                            type = "Class";
-                        }
-                        if (item.memberTypes.value.indexOf("Individual") > 0) {
-                            type = "Individual";
-                        }
-                        if (!existingIds[parentId]) {
-                            existingIds[parentId] = [];
-                        }
-                        var parentJstreeId = "_" + common.getRandomHexaId(5);
-                        existingIds[parentId].push(parentJstreeId);
-
-                        var id = item.member.value;
-                        if (!existingIds[id]) {
-                            existingIds[id] = [];
-                        }
-                        var jstreeId = "_" + common.getRandomHexaId(5);
-                        existingIds[id].push(jstreeId);
-
-                        // insert into into each parent
-                        var parents = existingIds[parentId];
-                        parents.forEach(function (jsTreeParent) {
-                            if (!existingNodes[jsTreeParent]) {
-                                existingNodes[jsTreeParent] = 1;
-                            }
-                            var node = {
-                                id: jstreeId,
-                                text: item.memberLabel.value,
-                                parent: jsTreeParent,
-                                type: type,
-                                data: {
-                                    type: type,
-                                    source: source,
-                                    id: id,
-                                    label: item.memberLabel.value,
-                                    currentParent: jsTreeParent,
-                                    tabId: options.tabId,
-                                },
-                            };
-
-                            jstreeData.push(node);
-                        });
-                        */
-                        var id = item.member.value;
-                        var jstreeId = "_" + common.getRandomHexaId(5);
-                        if (!existingIds[id]) {
-                            existingIds[id] = jstreeId;
-                        }
-                        var parentJstreeId = "_" + common.getRandomHexaId(5);
-                        if (!existingIds[parentId]) {
-                            existingIds[parentId] = parentJstreeId;
-                        }
-
-                        var node = {
-                            id: existingIds[id],
-                            text: item.memberLabel.value,
-                            parent: existingIds[parentId],
-                            type: type,
-                            data: {
-                                type: type,
-                                source: source,
-                                id: id,
-                                label: item.memberLabel.value,
-                                currentParent: existingIds[parentId],
-                                tabId: options.tabId
-                            }
-                        };
-
-                        jstreeData.push(node);
-                    });
-
-                    var jstreeOptions;
-                    if (options.jstreeOptions) {
-                        jstreeOptions = options.jstreeOptions;
-                    } else {
-                        jstreeOptions = {
-                            openAll: false,
-                            contextMenu: Lineage_containers.getContextJstreeMenu(),
-                            selectTreeNodeFn: Lineage_containers.onSelectedNodeTreeclick,
-                            dnd: {
-                                drag_stop: function(data, element, helper, event) {
-                                    //  self.onMoveContainer(data, element, helper, event);
-                                },
-                                drag_start: function(data, element, helper, event) {
-                                    var sourceNodeId = element.data.nodes[0];
-                                    self.currenDraggingNodeSourceParent = $("#lineage_containers_containersJstree").jstree().get_node(sourceNodeId).parent;
-                                }
-                            }
-                        };
-                    }
-
-                    JstreeWidget.loadJsTree(jstreeDiv, jstreeData, jstreeOptions, function() {
-                        if (options.filter) {
-                            $("#" + jstreeDiv).jstree("open_all");
-                        } else {
-                            $("#" + jstreeDiv)
-                                .jstree()
-                                .open_node("#");
-                        }
-
-                        self.bindMoveNode(jstreeDiv);
-                    });
-
-                    callbackSeries();
-                }
-            ],
-            function(err) {
+        var otherSourcesNodes = [];
+        var triples = [];
+        nodesData.forEach(function(nodeData) {
+            if (container.id == nodeData.id) {
+                return alert("a  node cannot be member of itself");
             }
-        );
+
+            if (!nodeData.source) {
+                return console.log(" node without source");
+            }
+            if (true || nodeData.source == source) {
+                triples.push({
+                    subject: "<" + container.data.id + ">",
+                    predicate: "<http://www.w3.org/2000/01/rdf-schema#member>",
+                    object: "<" + nodeData.id + ">"
+                });
+            } else {
+                otherSourcesNodes.push(node.id);
+            }
+        });
+
+        Sparql_generic.insertTriples(source, triples, null, function(err, result) {
+            if (err) {
+                if (callback) {
+                    return callback(err);
+                }
+                return alert(err.responseText);
+            }
+            MainController.UI.message("nodes added to container " + container.label);
+            var jstreeData = [];
+            nodesData.forEach(function(nodeData) {
+                jstreeData.push({
+                    id: nodeData.id + "_" + common.getRandomHexaId(5),
+                    text: nodeData.label,
+                    parent: container.id,
+                    type: "class",
+                    data: {
+                        type: nodesData.type || "resource",
+                        source: source,
+                        id: nodeData.id,
+                        label: nodeData.label
+                    }
+                });
+            });
+            if ($("#lineage_containers_containersJstree").jstree) {
+                if (callback) {
+                    JstreeWidget.addNodesToJstree("lineage_containers_containersJstree", container.id, jstreeData, null, callback);
+                } else {
+                    JstreeWidget.addNodesToJstree("lineage_containers_containersJstree", container.id, jstreeData);
+                }
+            }
+
+            if (drawMembershipEdge) {
+                var existingNodes = Lineage_whiteboard.lineageVisjsGraph.getExistingIdsMap();
+                var edges = [];
+                nodesData.forEach(function(nodeData) {
+                    var edgeId = container.id + "_" + "member" + "_" + nodeData.id;
+                    if (!existingNodes[edgeId]) {
+                        existingNodes[edgeId] = 1;
+                        edges.push({
+                            id: edgeId,
+                            from: container.id,
+                            to: nodeData.id,
+                            arrows: {
+                                from: {
+                                    enabled: true,
+                                    type: Lineage_whiteboard.defaultEdgeArrowType,
+                                    scaleFactor: 0.5,
+                                },
+                            },
+
+                            data: { propertyId: "http://www.w3.org/2000/01/rdf-schema#member", source: source },
+                            font: { multi: true, size: 10 },
+
+                            //  dashes: true,
+                            color: self.containerEdgeColor
+                        });
+                    }
+                });
+
+                Lineage_whiteboard.lineageVisjsGraph.data.edges.add(edges);
+            }
+            if (callback) {
+                return callback(null);
+                if (jstreeData) {
+                    callback(jstreeData);
+                }
+            }
+        });
     };
+
+    self.pasteNodesInContainer = function(source, container) {
+        common.pasteTextFromClipboard(function(text) {
+            // debugger
+            if (!text) {
+                return MainController.UI.message("no node copied");
+            }
+            try {
+                var nodes = JSON.parse(text);
+                var nodesData = [];
+                nodes.forEach(function(node) {
+                    if (node.data) {
+                        nodesData.push(node.data);
+                    }
+                });
+
+                self.addResourcesToContainer(source, container, nodesData);
+            } catch (e) {
+                console.log("wrong clipboard content");
+            }
+            return;
+        });
+    };
+
+    self.deleteContainer = function(source, container) {
+        if (!confirm("delete container)")) {
+            return;
+        }
+        self.currentContainer = null;
+        Sparql_generic.deleteTriples(source, container.data.id, null, null, function(err) {
+            if (err) {
+                return alert(err.responseText);
+            }
+
+            Sparql_generic.deleteTriples(source, null, null, container.data.id, function(err) {
+                var node = $("#lineage_containers_containersJstree").jstree().get_node(container.id);
+                if (node.children.length > 0) {
+                    $("#lineage_containers_containersJstree").jstree().move_node(node.children, "#");
+                }
+                $("#lineage_containers_containersJstree").jstree().delete_node(container.id);
+            });
+        });
+    };
+    /**
+     *
+     *
+     *
+     *
+     * @param source
+     * @param containerIds
+     * @param options
+     *  allDescendants
+     *  nodes | containers
+     *  descendants
+     *
+     * @param callback
+     */
 
     self.bindMoveNode = function(jstreeDiv) {
         $("#" + jstreeDiv).bind("move_node.jstree", function(e, data) {
@@ -448,240 +323,6 @@ var Lineage_containers = (function() {
         });
     };
 
-    /**
-     *
-     * add nodes to a container(owl:bag)
-     *
-     *
-     * @param source
-     * @param container
-     * @param nodesData
-     * @param drawMembershipEdge add the edge (and the node) on the vizGraph
-     */
-    self.addResourcesToContainer = function(source, container, nodesData, drawMembershipEdge, callback) {
-        if (!(container.data.type.includes("http://www.w3.org/1999/02/22-rdf-syntax-ns#Bag") || container.data.type == "container")) {
-            return alert("can only add resources to containers");
-        }
-        // self.currentContainer=null;
-        if (!Array.isArray(nodesData)) {
-            nodesData = [nodesData];
-        }
-
-        var otherSourcesNodes = [];
-        var triples = [];
-        nodesData.forEach(function(nodeData) {
-            if (container.id == nodeData.id) {
-                return alert("a  node cannot be member of itself");
-            }
-
-            if (!nodeData.source) {
-                return console.log(" node without source");
-            }
-            if (true || nodeData.source == source) {
-                triples.push({
-                    subject: "<" + container.data.id + ">",
-                    predicate: "<http://www.w3.org/2000/01/rdf-schema#member>",
-                    object: "<" + nodeData.id + ">"
-                });
-            } else {
-                otherSourcesNodes.push(node.id);
-            }
-        });
-
-        Sparql_generic.insertTriples(source, triples, null, function(err, result) {
-            if (err) {
-                if (callback) {
-                    return callback(err);
-                }
-                return alert(err.responseText);
-            }
-            MainController.UI.message("nodes added to container " + container.label);
-            var jstreeData = [];
-            nodesData.forEach(function(nodeData) {
-                jstreeData.push({
-                    id: nodeData.id + "_" + common.getRandomHexaId(5),
-                    text: nodeData.label,
-                    parent: container.id,
-                    type: "class",
-                    data: {
-                        type: nodesData.type || "resource",
-                        source: source,
-                        id: nodeData.id,
-                        label: nodeData.label
-                    }
-                });
-            });
-            if ($("#lineage_containers_containersJstree").jstree) {
-                if (callback) {
-                    JstreeWidget.addNodesToJstree("lineage_containers_containersJstree", container.id, jstreeData, null, callback);
-                } else {
-                    JstreeWidget.addNodesToJstree("lineage_containers_containersJstree", container.id, jstreeData);
-                }
-            }
-
-            if (drawMembershipEdge) {
-                var existingNodes = Lineage_whiteboard.lineageVisjsGraph.getExistingIdsMap();
-                var edges = [];
-                nodesData.forEach(function(nodeData) {
-                    var edgeId = container.id + "_" + "member" + "_" + nodeData.id;
-                    if (!existingNodes[edgeId]) {
-                        existingNodes[edgeId] = 1;
-                        edges.push({
-                            id: edgeId,
-                            from: container.id,
-                            to: nodeData.id,
-                           // arrows: " middle",
-
-                            data: { propertyId: "http://www.w3.org/2000/01/rdf-schema#member", source: source },
-                            font: { multi: true, size: 10 },
-
-                            //  dashes: true,
-                            color: self.containerEdgeColor
-                        });
-                    }
-                });
-
-                Lineage_whiteboard.lineageVisjsGraph.data.edges.add(edges);
-            }
-            if (callback) {
-                return callback(null);
-                if (jstreeData) {
-                    callback(jstreeData);
-                }
-            }
-        });
-    };
-
-    self.pasteNodesInContainer = function(source, container) {
-        common.pasteTextFromClipboard(function(text) {
-            // debugger
-            if (!text) {
-                return MainController.UI.message("no node copied");
-            }
-            try {
-                var nodes = JSON.parse(text);
-                var nodesData = [];
-                nodes.forEach(function(node) {
-                    if (node.data) {
-                        nodesData.push(node.data);
-                    }
-                });
-
-                self.addResourcesToContainer(source, container, nodesData);
-            } catch (e) {
-                console.log("wrong clipboard content");
-            }
-            return;
-        });
-    };
-
-    self.deleteContainer = function(source, container) {
-        if (!confirm("delete container)")) {
-            return;
-        }
-        self.currentContainer = null;
-        Sparql_generic.deleteTriples(source, container.data.id, null, null, function(err) {
-            if (err) {
-                return alert(err.responseText);
-            }
-
-            Sparql_generic.deleteTriples(source, null, null, container.data.id, function(err) {
-                var node = $("#lineage_containers_containersJstree").jstree().get_node(container.id);
-                if (node.children.length > 0) {
-                    $("#lineage_containers_containersJstree").jstree().move_node(node.children, "#");
-                }
-                $("#lineage_containers_containersJstree").jstree().delete_node(container.id);
-            });
-        });
-    };
-    /**
-     *
-     *
-     *
-     *
-     * @param source
-     * @param containerIds
-     * @param options
-     *  allDescendants
-     *  nodes | containers
-     *  descendants
-     *
-     * @param callback
-     */
-
-    self.listContainerResources = function(source, containerNode, options, callback, JstreeDiv) {
-        var existingChildren = [];
-        if (!JstreeDiv) {
-            var JstreeDiv = "lineage_containers_containersJstree";
-        }
-
-        //Do not execute the descendants request if he already has children?
-        if (containerNode.children.length > 1) {
-            return;
-        }
-
-        self.sparql_queries.getContainerDescendants(source, containerNode ? containerNode.data.id : null, options, function(err, result) {
-            if (err) {
-                return alert(err.responseText);
-            }
-
-            var existingNodes = {};
-            if (containerNode) {
-                // existingNodes=$("#lineage_containers_containersJstree").jstree().get_node(containerNode.id).children;
-                var jstreeChildren = JstreeWidget.getNodeDescendants(JstreeDiv, containerNode.id, 2);
-                jstreeChildren.forEach(function(item) {
-                    existingNodes[item.data.id] = 1;
-                });
-            }
-
-            var jstreeData = [];
-            var nodesMap = {};
-
-            result.results.bindings.forEach(function(item) {
-                //  var nodeId=item.parent+"_"+item.member.value
-                item.jstreeId = "_" + common.getRandomHexaId(5);
-                nodesMap[item.member.value] = item;
-            });
-
-            for (var key in nodesMap) {
-                var item = nodesMap[key];
-
-                var containerJstreeId = "#";
-                var continerDataId = null;
-                if (containerNode) {
-                    containerJstreeId = containerNode.id;
-                    continerDataId = containerNode.data.id;
-                }
-
-                var parent = item.parent.value == continerDataId ? containerJstreeId : nodesMap[item.parent.value] ? nodesMap[item.parent.value].jstreeId : "#";
-                if (!existingNodes[item.member.value]) {
-                    existingNodes[item.member.value] = 1;
-
-                    var types = item.memberTypes.value.split(",");
-                    var type_icon = JstreeWidget.selectTypeForIconsJstree(types, self.add_supplementary_layer_for_types_icon);
-
-                    jstreeData.push({
-                        id: item.jstreeId,
-                        text: item.memberLabel.value,
-                        parent: parent,
-                        type: type_icon,
-                        data: {
-                            type: types,
-                            source: source,
-                            id: item.member.value,
-                            label: item.memberLabel.value
-                        }
-                    });
-                }
-            }
-            if (jstreeData.length > 0) {
-                JstreeWidget.addNodesToJstree(JstreeDiv, containerJstreeId, jstreeData);
-                if (callback) {
-                    callback(jstreeData);
-                }
-            }
-        });
-    };
 
     self.graphResources = function(source, containerData, options, callback) {
         if (!options) {
@@ -738,10 +379,10 @@ var Lineage_containers = (function() {
                             id: containerData.id,
                             label: containerData.label,
                             shadow: self.nodeShadow,
-                            shape: Lineage_containers.containerStyle.shape,
+                            shape: Lineage_containers_UI.containerStyle.shape,
                             size: size,
                             font: type == "container" ? { color: "#70309f" } : null,
-                            color: Lineage_containers.containerStyle.color,
+                            color: Lineage_containers_UI.containerStyle.color,
                             data: {
                                 type: type,
                                 source: source,
@@ -759,10 +400,10 @@ var Lineage_containers = (function() {
                                 id: item.parent.value,
                                 label: item.parentLabel.value,
                                 shadow: self.nodeShadow,
-                                shape: type == "container" ? Lineage_containers.containerStyle.shape : shape,
+                                shape: type == "container" ? Lineage_containers_UI.containerStyle.shape : shape,
                                 size: size,
                                 font: type == "container" ? { color: color2, size: 10 } : null,
-                                color: Lineage_containers.containerStyle.color,
+                                color: Lineage_containers_UI.containerStyle.color,
                                 data: {
                                     type: type,
                                     source: source,
@@ -773,8 +414,8 @@ var Lineage_containers = (function() {
                         }
 
                         if (!existingNodes[item.member.value]) {
-                            var color = Lineage_containers.containerStyle.color;
-                            var shape = Lineage_containers.containerStyle.shape;
+                            var color = Lineage_containers_UI.containerStyle.color;
+                            var shape = Lineage_containers_UI.containerStyle.shape;
                             var type = "container";
                             if (item.memberTypes.value.indexOf("Bag") < 0) {
                                 color = Lineage_whiteboard.getSourceColor(Lineage_sources.activeSource);
@@ -817,13 +458,13 @@ var Lineage_containers = (function() {
                                     id: edgeId,
                                     from: containerData.id,
                                     to: item.parent.value,
-                                  /*  arrows: {
-                                        middle: {
+                                    arrows: {
+                                        from: {
                                             enabled: true,
                                             type: Lineage_whiteboard.defaultEdgeArrowType,
                                             scaleFactor: 0.5
                                         }
-                                    },*/
+                                    },
                                     data: {
                                         from: containerData.id,
                                         to: item.parent.value,
@@ -941,13 +582,7 @@ var Lineage_containers = (function() {
         );
     };
 
-    self.onSelectedNodeTreeclick = function(event, obj) {
-        self.currentContainer = obj.node;
 
-        if (obj.event.button != 2) {
-            self.listContainerResources(Lineage_sources.activeSource, self.currentContainer, { onlyOneLevel: true, leaves: true });
-        }
-    };
 
     self.writeMovedNodeNewParent = function(movedNodeInfos) {
         var graphUri = Config.sources[Lineage_sources.activeSource].graphUri;
@@ -975,93 +610,7 @@ var Lineage_containers = (function() {
         });
     };
 
-    self.sparql_queries = {
-        getTopContainers: function(source, callback) {
-            var fromStr = Sparql_common.getFromStr(source, false, false);
-            var query =
-                "PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>\n" +
-                "PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>\n" +
-                "SELECT distinct ?member ?memberLabel " +
-                fromStr +
-                " where {" +
-                "    ?member rdf:type ?memberType. " +
-                " ?member rdfs:label ?memberLabel. " +
-                " FILTER (?memberType in(rdf:Bag,rdf:List))\n" +
-                "  filter (not exists{?parent rdfs:member ?member})\n" +
-                "    }";
 
-            var url = Config.sources[source].sparql_server.url + "?format=json&query=";
-
-            Sparql_proxy.querySPARQL_GET_proxy(url, query, "", { source: source }, function(err, result) {
-                if (err) {
-                    return callback(err);
-                }
-                return callback(null, result);
-            });
-        },
-
-        getContainerDescendants: function(source, containerId, options, callback) {
-            var fromStr = Sparql_common.getFromStr(source, false, false);
-            var filterContainer0Str = "";
-            if (containerId) {
-                // needs options.useFilterKeyWord because VALUES dont work
-                filterContainer0Str = Sparql_common.setFilter("parent0", containerId, null, { useFilterKeyWord: 1 });
-            }
-
-            var filter = options.filter || "";
-
-            var filterLeaves = "";
-            if (!options.leaves) {
-                filterLeaves = " FILTER (?memberType in(rdf:Bag,rdf:List))";
-            }
-
-            //  var pathOperator = "+";
-            var pathOperator = "+";
-            if (options.onlyOneLevel) {
-                pathOperator = "";
-            } else if (options.depth) {
-                pathOperator = "{1," + options.depth + "}";
-            }
-            var query = `PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#> 
-                 PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#> 
-                 SELECT distinct ?member ?memberLabel ?parent ?parentLabel (GROUP_CONCAT( distinct ?memberType;separator=",") as ?memberTypes)  ${fromStr}
-                 WHERE {?searchValue ^rdfs:member* ?member.
-                    ?member ^rdfs:member ?parent.
-                    ?member rdf:type ?memberType.
-                    ?member rdfs:label ?memberLabel.
-                    ?parent rdfs:label ?parentLabel.
-                    
-                    {select ?searchValue where{
-                        ?parent0  rdfs:member${pathOperator} ?searchValue.
-                        ?searchValue rdfs:label ?searchValueLabel.
-                        ${filterContainer0Str}
-                        ${filter}
-                        }
-                    }
-                }  group by   ?member ?memberLabel ?parent ?parentLabel ?searchValue   
-                `;
-
-            /*
-                filter +
-                "  {select ?member where{\n" +
-                "?parent0  rdfs:member" +
-                pathOperator +
-                " ?member." +
-                filterContainer0Str +
-                "}\n" +
-                "  }\n" +
-                "}  group by ?member ?memberLabel ?parent ?parentLabel";
-                    */
-            var url = Config.sources[source].sparql_server.url + "?format=json&query=";
-
-            Sparql_proxy.querySPARQL_GET_proxy(url, query, "", { source: source }, function(err, result) {
-                if (err) {
-                    return callback(err);
-                }
-                return callback(null, result);
-            });
-        }
-    };
 
     self.applyContainerstyle = function(containerUrl) {
     };
@@ -1114,10 +663,10 @@ var Lineage_containers = (function() {
                         id: item.container.value,
                         label: item.containerLabel.value,
                         shadow: self.nodeShadow,
-                        shape: Lineage_containers.containerStyle.shape,
+                        shape: Lineage_containers_UI.containerStyle.shape,
                         size: Lineage_whiteboard.defaultShapeSize,
                         font: { color: color2 },
-                        color: Lineage_containers.containerStyle.color,
+                        color: Lineage_containers_UI.containerStyle.color,
                         data: {
                             type: "container",
                             source: Lineage_sources.activeSource,
@@ -1135,9 +684,16 @@ var Lineage_containers = (function() {
                         id: edgeId,
                         from: item.container.value,
                         to: item.node.value,
-                    //    arrows: "middle",
+                      /*  arrows: {
+                            to: {
+                                enabled: true,
+                                type: Lineage_whiteboard.defaultEdgeArrowType,
+                                scaleFactor: 0.5,
+                            },
+                        },*/
+
                         data: { from: item.container.value, to: item.node.value, source: source },
-                        font: { multi: true, size: 10 },
+                        //  font: { multi: true, size: 10 },
 
                         //  dashes: true,
                         color: self.containerEdgeColor
@@ -1187,10 +743,9 @@ var Lineage_containers = (function() {
             return callback(null, types);
         });
     };
-
     return self;
 })();
 
-export default Lineage_containers;
+export default Containers_UI;
 
-window.Lineage_containers = Lineage_containers;
+window.Containers_UI = Containers_UI;
